--- alembic/versions/3e82f78503da_add_memorystat.py
@@ -1,34 +1,0 @@
-"""add MemoryStat
-
-Revision ID: 3e82f78503da
-Revises: 3d3239ae4849
-Create Date: 2023-06-22 15:59:02.998365
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '3e82f78503da'
-down_revision = '3d3239ae4849'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('memory_stat',
-                    sa.Column('id', sa.Integer(), nullable=False),
-                    sa.Column('recording_timestamp', sa.Integer(), nullable=True),
-                    sa.Column('memory_usage', sa.Integer(), nullable=True),
-                    sa.Column('timestamp', sa.Integer(), nullable=True),
-                    sa.PrimaryKeyConstraint('id', name='pk_memory_stat')
-                    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('memory_stat')
-    # ### end Alembic commands ###
