+++ alembic/versions/607d1380b5ae_add_memorystat.py
@@ -1,0 +1,35 @@
+"""add MemoryStat
+
+Revision ID: 607d1380b5ae
+Revises: 104d4a614d95
+Create Date: 2023-06-28 11:54:36.749072
+
+"""
+from alembic import op
+import sqlalchemy as sa
+from openadapt.models import ForceFloat
+
+
+# revision identifiers, used by Alembic.
+revision = '607d1380b5ae'
+down_revision = '104d4a614d95'
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('memory_stat',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('recording_timestamp', sa.Integer(), nullable=True),
+    sa.Column('memory_usage_bytes', ForceFloat(precision=10, scale=2, asdecimal=False), nullable=True),
+    sa.Column('timestamp', ForceFloat(precision=10, scale=2, asdecimal=False), nullable=True),
+    sa.PrimaryKeyConstraint('id', name=op.f('pk_memory_stat'))
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('memory_stat')
+    # ### end Alembic commands ###
